backend/requirements.txt
fastapi
uvicorn[standard]
devtools
requests
pywebview
annotated-types
Pillow
opencv-python-headless
cachetools
python-multipart
redis
pytest
pytest-asyncio
tuna